move fs and types out to folder
