+++ src/lib/terminal/index.ts
@@ -1,0 +1,2 @@
+export * from "./filesystem";
+export * from "./types";
