+++ src/lib/views/InteractiveTerminalDom.svelte
@@ -1,138 +1,8 @@
 <script lang="ts">
+    import type {Command, ExecContext, Line, Segment, ExecResult, HistoryEntry, PromptSnapshot, DirNode, FSNode} from "$lib/terminal";
+    import {cwdAbs, cwdString, getNode, makeFilesystem, resolveParts} from "$lib/terminal/filesystem";
     import {tick} from "svelte";
 
-    // ── Types ──────────────────────────────────────────────────────────────────
-
-    type FileNode = {type: "file"; content: string; executable?: boolean};
-    type DirNode = {type: "dir"; children: Record<string, FSNode>};
-    type FSNode = FileNode | DirNode;
-
-    type Segment = {
-        text: string;
-        palette?: number; // index into --config-palette-N
-        hex?: string; // explicit hex color (for git, grep highlights, etc.)
-        bold?: boolean;
-        dim?: boolean;
-        italic?: boolean;
-    };
-
-    type Line = Segment[];
-
-    type ExecResult = {lines: Line[]; clear?: boolean; mutateCwd?: string[]};
-
-    type ExecContext = {
-        cwd: () => string[]; // getter so commands always see latest
-        setCwd: (p: string[]) => void;
-        root: DirNode;
-        user: string;
-        host: string;
-    };
-
-    type Command = {
-        desc: string; // one-liner shown in `help`
-        usage?: string; // args shown in `help`, e.g. "[-la] [path]"
-        details?: string[]; // extra lines shown in `help <cmd>`
-        fn: (args: string[], ctx: ExecContext) => ExecResult;
-    };
-
-    type HistoryEntry =
-        | {kind: "cmd"; prompt: PromptSnapshot; cmd: string}
-        | {kind: "output"; id: number; segments: Segment[]}
-        | {kind: "ctrlc"; prompt: PromptSnapshot; input: string};
-
-    type PromptSnapshot = {user: string; host: string; cwd: string};
-
-    // ── Filesystem ─────────────────────────────────────────────────────────────
-
-    function makeFilesystem(): DirNode {
-        return {
-            type: "dir",
-            children: {
-                "Desktop": {type: "dir", children: {}},
-                "Documents": {
-                    type: "dir",
-                    children: {
-                        "notes.txt": {type: "file", content: "Meeting notes – Jan 1\nTODO: update ghostty config"},
-                        "todo.md": {type: "file", content: "# TODO\n- [ ] Configure ghostty\n- [x] Install ghostty\n- [ ] Customize colors"},
-                        "report.pdf": {type: "file", content: "(binary file)"},
-                    },
-                },
-                "Downloads": {
-                    type: "dir",
-                    children: {
-                        "ghostty-1.0.0.tar.gz": {type: "file", content: "(binary file)"},
-                        "setup.sh": {type: "file", content: "#!/bin/bash\napt-get install -y ghostty", executable: true},
-                    },
-                },
-                "Pictures": {
-                    type: "dir",
-                    children: {
-                        "screenshot.png": {type: "file", content: "(binary file)"},
-                        "wallpaper.jpg": {type: "file", content: "(binary file)"},
-                    },
-                },
-                "Projects": {
-                    type: "dir",
-                    children: {
-                        "ghostty-config": {
-                            type: "dir",
-                            children: {
-                                "README.md": {type: "file", content: "# ghostty-config\nA web-based GUI config generator for Ghostty."},
-                                "package.json": {type: "file", content: "{\n  \"name\": \"ghostty-config\",\n  \"version\": \"1.0.0\"\n}"},
-                                "src": {
-                                    type: "dir",
-                                    children: {
-                                        "index.ts": {type: "file", content: "import \"./app.css\";\nimport App from \"./App.svelte\";"},
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-                ".bash_profile": {type: "file", content: "export PATH=\"$HOME/.local/bin:$PATH\"\nexport TERM=ghostty"},
-                ".gitconfig": {type: "file", content: "[user]\n\tname = John\n\temail = john@example.com\n[core]\n\teditor = nvim"},
-                "README.md": {type: "file", content: "# My Ghostty Setup\n\nA customized Ghostty configuration.\nGenerated with ghostty-config."},
-                "install.sh": {type: "file", content: "#!/bin/bash\necho \"Installing...\"\napt-get install -y ghostty", executable: true},
-            },
-        };
-    }
-
-    // ── Path helpers ───────────────────────────────────────────────────────────
-
-    function resolveParts(cwd: string[], path: string): string[] {
-        let parts: string[];
-        if (path === "~" || path === "") parts = [];
-        else if (path.startsWith("~/")) parts = path.slice(2).split("/").filter(Boolean);
-        else if (path.startsWith("/")) parts = path.slice(1).split("/").filter(Boolean);
-        else parts = [...cwd, ...path.split("/").filter(Boolean)];
-
-        const out: string[] = [];
-        for (const p of parts) {
-            if (p === ".") continue;
-            else if (p === "..") out.pop();
-            else out.push(p);
-        }
-        return out;
-    }
-
-    function getNode(root: DirNode, parts: string[]): FSNode | null {
-        let node: FSNode = root;
-        for (const part of parts) {
-            if (node.type !== "dir") return null;
-            const child = node.children[part] as FSNode | undefined;
-            if (!child) return null;
-            node = child;
-        }
-        return node;
-    }
-
-    function cwdString(parts: string[]): string {
-        return parts.length === 0 ? "~" : "~/" + parts.join("/");
-    }
-
-    function cwdAbs(user: string, parts: string[]): string {
-        return parts.length === 0 ? `/home/${user}` : `/home/${user}/${parts.join("/")}`;
-    }
 
     // ── Segment helpers ────────────────────────────────────────────────────────
